backend/requirements.txt
attrs==19.3.0
certifi>=2020.6.20
decorator==4.4.1
flask==1.1.1
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.5.0
marshmallow==3.4.0
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
python-dotenv==0.14.0
pyparsing==2.4.6
pytest-cov==2.10.0
pytest-flask==1.0.0
pytest-mock==3.2.0
pytest-watch>=4.2.0
six==1.14.0
sqreen==1.21.1
wcwidth==0.1.8
zipp==2.1.0
